+++ endpoint.go
@@ -2,21 +2,10 @@ package p2p
 
 import (
 	"fmt"
-	math "math"
-	"math/rand"
 	"net"
 	"regexp"
 	"strconv"
 )
-
-func testRandomEndpoint() Endpoint {
-	ip := fmt.Sprintf("%d.%d.%d.%d", 1+rand.Intn(255), rand.Intn(256), rand.Intn(256), 1+rand.Intn(255))
-	return Endpoint{IP: ip, Port: testRandomPort()}
-}
-
-func testRandomPort() string {
-	return fmt.Sprintf("%d", (1+rand.Uint32())%math.MaxUint16)
-}
 
 var hostnameRegex *regexp.Regexp
