+++ prometheus.go
@@ -75,8 +75,8 @@ func (p *Prometheus) Setup() {
 		p.MessageRateUp = ng("factomd_p2p_msg_rate_up", "Current rate of outgoing traffic (in messages/sec)")
 
 		p.DroppedPeerSend = ng("factomd_p2p_dropped_peer_send", "Number of parcels that were unable to be sent because of peer's lack of bandwidth")
-		p.DroppedFromNetwork = ng("factomd_p2p_dropped_from_network", "Number of parcels that were dropped because application is sending parcels too fast")
-		p.DroppedToNetwork = ng("factomd_p2p_dropped_to_network", "Number of parcels that were dropped because application is not reading parcels fast enough")
+		p.DroppedFromNetwork = ng("factomd_p2p_dropped_from_network", "Number of parcels that were dropped because application is not reading parcels fast enough")
+		p.DroppedToNetwork = ng("factomd_p2p_dropped_to_network", "Number of parcels that were dropped because application is sending parcels too fast")
 
 		p.ParcelSize = prometheus.NewHistogram(prometheus.HistogramOpts{
 			Name:    "factomd_p2p_parcels_size",
